requirements.txt
langchain>=0.2.11
langgraph>=0.2.2
langchain-google-genai>=1.0.8
python-dotenv>=1.0.1
requests>=2.32.3
pydantic>=2.8.2
